finished tpu model runner
